requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
assemblyai==0.23.1
attrs==23.2.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.1
g4f==0.2.0.7
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
groq==0.4.2
grpcio==1.62.1
grpcio-status==1.62.1
gruut_lang_de==2.0.0
gruut_lang_es==2.0.0
gruut_lang_fr==2.0.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.25.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mistralai==0.1.6
moviepy==1.0.3
multidict==6.0.5
numpy==1.26.4
oauth2client==4.1.3
openai==1.14.2
orjson==3.9.15
outcome==1.3.0.post0
pandas==1.5.3
Pillow==9.5.0
pip-autoremove==0.10.0
playsound==1.3.0
proglog==0.1.10
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.26.0
pydantic==2.6.4
pydantic_core==2.16.3
PyGObject==3.48.1
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
rsa==4.9
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
srt==3.5.3
srt_equalizer==0.1.8
SudachiDict-core==20240109
SudachiPy==0.6.8
termcolor==2.4.0
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.4
